refactor: move avaliacoes tab to its own directory
