+++ src/screens/private/client/Profile/Tabs/AvaliacoesTab/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AvaliacoesTab';
